service method 관련 로딩을 ServiceDirectory 로 위임
